Add EventExtent represnting an extent of numbers and durations of events
The window length of the remembered metadata per dstream has became
dynamic. In addition to a fixed duration, metadata could be kept for
a number of events, as events are not necessarily eqi-spaced in time
, the remember window duration varies with time.

Moreover, some dstreams could base the remember interval on time
and some on counting events. Parent streams are set to remember a
sequence of remember intervals from its children. Evaluating the
sequence of those remember intervals needs to be done dynamically.

EventExtent keeps a DAG of intervals resembling the interval each
dstream remembers up to the holding dstream.